enhancement (try 2): these generators now check if there is already a class on the classpath that matches the decorator/plugin (instead of only handwritten/java).
This allows to for example export gcc project to a jar and reuse its plugins in another project (use case: using auth outside of gcc). Or to seperate out the code over multiple source folders. Caveat is that the .ftl is assumed to be there if the java is there. Future todo is also prevend duplication of the generated code.

git-svn-id: http://www.molgenis.org/svn/molgenis/trunk@7125 42de9c89-b41c-0410-9fa9-f33e7b6bda7a