Removed isVisible() from GenericPlugin since that extends from PluginModel so it already has it.
Added 'canRead' check to isVisible(), so plugins are only shown to people who have read rights on them (and to superusers).

Regarding User story #833:
I tested this with groups and it works perfectly. Create a group, add some users and give the group read (or higher) rights on the app.ui.PluginInQuestion. If you want to have these rights pre-made for you, add the corresponding code to your (Pre)Filldatabase plugin.

git-svn-id: http://www.molgenis.org/svn/molgenis_apps/trunk@7650 42de9c89-b41c-0410-9fa9-f33e7b6bda7a