Removed useless files.
Got ListPlugin working again on VM7. Some essential AnimalDB-specific code which was in UserLogin was brutally commented out WITHOUT notifying me. The code was causing unwanted dependencies, that's true, but an email would've been nice!!
Now solved it by moving the code to the ListPlugin itself and introducing a workaround with an extra user request to get it working.

git-svn-id: http://www.molgenis.org/svn/molgenis_apps/trunk@8455 42de9c89-b41c-0410-9fa9-f33e7b6bda7a