ftl files for compute should now contain as header: #MOLGENIS walltime=08:00:00 nodes=1 cores=4 mem=6 target=lane
All these fields are in ComputeProtocol.walltime,nodes,cores,mem,targetType

We shouldn't use #PBS because then users expect all PBS commands to work. This ensure that we have a clear 'contract'. 
Discussion: we we want users to still able to add #PBS commands later in the file if they like?

M

git-svn-id: http://www.molgenis.org/svn/molgenis_apps/trunk@9528 42de9c89-b41c-0410-9fa9-f33e7b6bda7a