fixed selenium test to handel the changes in the experiments plugin (this should fix the build).
Although selenium is happy again this does not mean that everything works. When editing existing decs and experiments the changes are not saved (althoug the happy green bar is shown)
I wonder wheter this was actually working before. Judging by the comments in the code this was previously relying  on the getmost recent ... value mechanism, anticipating that a change would be saved as a new measurment/value combi.

I think we should change this to updating existing values in stead.

git-svn-id: http://www.molgenis.org/svn/molgenis_apps/trunk@9663 42de9c89-b41c-0410-9fa9-f33e7b6bda7a